requirements.txt
fastapi
uvicorn
vonage
python-dotenv
openai
vonage>=4.0.0
